fix(ARC-2077): fix issues with admin-core routes in english
