+++ src/modules/shared/const/routes.ts
@@ -18,7 +18,7 @@ export enum RoutePart {
 	translations = 'translations',
 	admin = 'admin',
 	cpAdmin = 'cpAdmin',
-	content = 'content',
+	contentPages = 'contentPages',
 	contentPageLabels = 'contentPageLabels',
 	cookiePolicy = 'cookiePolicy',
 	create = 'create',
@@ -53,7 +53,7 @@ const ROUTE_PARTS_NL: Record<RoutePart, string> = {
 	activeVisitors: 'actieve-bezoekers',
 	admin: 'admin',
 	alerts: 'meldingen',
-	content: 'content',
+	contentPages: 'content-paginas',
 	contentPageLabels: 'content-pagina-labels',
 	cookiePolicy: 'cookiebeleid',
 	cpAdmin: 'beheer',
@@ -97,7 +97,7 @@ const ROUTE_PARTS_EN: Record<RoutePart, string> = {
 	activeVisitors: 'active-visitors',
 	admin: 'admin',
 	alerts: 'alerts',
-	content: 'content',
+	contentPages: 'content-pages',
 	contentPageLabels: 'content-page-labels',
 	cookiePolicy: 'cookiepolicy',
 	cpAdmin: 'management',
@@ -221,10 +221,10 @@ const ROUTES_NL: Record<RouteKey, string> = {
 	admin: `/${ROUTE_PARTS_NL.admin}`,
 	adminActiveVisitors: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.visitorSpaceManagement}/${ROUTE_PARTS_NL.activeVisitors}`,
 	adminAlerts: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.alerts}`,
-	adminContentPageCreate: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.content}/${ROUTE_PARTS_NL.create}`,
+	adminContentPageCreate: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPages}/${ROUTE_PARTS_NL.create}`,
 	adminContentPageLabelCreate: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPageLabels}/${ROUTE_PARTS_NL.create}`,
 	adminContentPageLabels: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPageLabels}`,
-	adminContentPages: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.content}`,
+	adminContentPages: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPages}`,
 	adminMaterialRequests: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.materialRequests}`,
 	adminNavigation: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.navigation}`,
 	adminNavigationCreate: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.navigation}/${ROUTE_PARTS_NL.create}`,
@@ -265,10 +265,10 @@ const ROUTES_EN: Record<RouteKey, string> = {
 	admin: `/${ROUTE_PARTS_EN.admin}`,
 	adminActiveVisitors: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.visitorSpaceManagement}/${ROUTE_PARTS_EN.activeVisitors}`,
 	adminAlerts: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.alerts}`,
-	adminContentPageCreate: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.content}/${ROUTE_PARTS_EN.create}`,
+	adminContentPageCreate: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPages}/${ROUTE_PARTS_EN.create}`,
 	adminContentPageLabelCreate: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPageLabels}/${ROUTE_PARTS_EN.create}`,
 	adminContentPageLabels: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPageLabels}`,
-	adminContentPages: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.content}`,
+	adminContentPages: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPages}`,
 	adminMaterialRequests: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.materialRequests}`,
 	adminNavigation: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.navigation}`,
 	adminNavigationCreate: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.navigation}/${ROUTE_PARTS_EN.create}`,
@@ -301,10 +301,10 @@ const ROUTES_EN: Record<RouteKey, string> = {
 
 const ADMIN_CORE_ROUTES_NL: AdminConfig['routes'] = {
 	ADMIN_ALERTS_OVERVIEW: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.alerts}`,
-	ADMIN_CONTENT_PAGE_CREATE: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.content}/${ROUTE_PARTS_NL.create}`,
-	ADMIN_CONTENT_PAGE_DETAIL: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.content}/:id`,
-	ADMIN_CONTENT_PAGE_EDIT: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.content}/:id/${ROUTE_PARTS_NL.edit}`,
-	ADMIN_CONTENT_PAGE_OVERVIEW: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.content}`,
+	ADMIN_CONTENT_PAGE_CREATE: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPages}/${ROUTE_PARTS_NL.create}`,
+	ADMIN_CONTENT_PAGE_DETAIL: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPages}/:id`,
+	ADMIN_CONTENT_PAGE_EDIT: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPages}/:id/${ROUTE_PARTS_NL.edit}`,
+	ADMIN_CONTENT_PAGE_OVERVIEW: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPages}`,
 	ADMIN_CONTENT_PAGE_LABEL_CREATE: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPageLabels}/${ROUTE_PARTS_NL.create}`,
 	ADMIN_CONTENT_PAGE_LABEL_DETAIL: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPageLabels}/:id`,
 	ADMIN_CONTENT_PAGE_LABEL_EDIT: `/${ROUTE_PARTS_NL.admin}/${ROUTE_PARTS_NL.contentPageLabels}/:id/${ROUTE_PARTS_NL.edit}`,
@@ -327,10 +327,10 @@ const ADMIN_CORE_ROUTES_NL: AdminConfig['routes'] = {
 
 const ADMIN_CORE_ROUTES_EN: AdminConfig['routes'] = {
 	ADMIN_ALERTS_OVERVIEW: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.alerts}`,
-	ADMIN_CONTENT_PAGE_CREATE: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.content}/${ROUTE_PARTS_EN.create}`,
-	ADMIN_CONTENT_PAGE_DETAIL: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.content}/:id`,
-	ADMIN_CONTENT_PAGE_EDIT: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.content}/:id/${ROUTE_PARTS_EN.edit}`,
-	ADMIN_CONTENT_PAGE_OVERVIEW: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.content}`,
+	ADMIN_CONTENT_PAGE_CREATE: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPages}/${ROUTE_PARTS_EN.create}`,
+	ADMIN_CONTENT_PAGE_DETAIL: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPages}/:id`,
+	ADMIN_CONTENT_PAGE_EDIT: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPages}/:id/${ROUTE_PARTS_EN.edit}`,
+	ADMIN_CONTENT_PAGE_OVERVIEW: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPages}`,
 	ADMIN_CONTENT_PAGE_LABEL_CREATE: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPageLabels}/${ROUTE_PARTS_EN.create}`,
 	ADMIN_CONTENT_PAGE_LABEL_DETAIL: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPageLabels}/:id`,
 	ADMIN_CONTENT_PAGE_LABEL_EDIT: `/${ROUTE_PARTS_EN.admin}/${ROUTE_PARTS_EN.contentPageLabels}/:id/${ROUTE_PARTS_EN.edit}`,
